training_data_bot/requirements.txt
# Core dependencies for Training Data Bot
# ==========================================

# Session 1 - Core models and exceptions
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Session 2 - Configuration and logging
python-dotenv>=1.0.0
PyYAML>=6.0

# Session 3-4 - Document loading
httpx>=0.24.0                    # Async HTTP client for web loading
beautifulsoup4>=4.12.0          # HTML parsing
PyMuPDF>=1.23.0                 # PDF text extraction (fitz)
python-docx>=1.0.0              # Word document reading
aiofiles>=23.0.0                # Async file operations (optional)

# Session 5 - Text preprocessing
# nltk>=3.8                      # Not required for our implementation
# spacy>=3.6.0                   # Not required for our implementation

# Session 6 - AI integration
openai>=1.0.0                   # OpenAI API client
anthropic>=0.25.0               # Anthropic API client (updated version)
tiktoken>=0.5.0                 # Token counting for OpenAI

# Session 7 - Task processing
# jinja2>=3.1.0                  # Not required - we use f-strings

# Session 8 - Storage and export
pandas>=2.0.0                   # Data manipulation for exports
pyarrow>=13.0.0                 # Parquet format support

# Development and testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0                   # Code formatting

python-dotenv>=1.0.0